safari-600.7-branch/LayoutTests/http/tests/xmlhttprequest/access-control-basic-denied-preflight-cache-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
Test async xhr preflight cache denial. If this test passes, there should be a single PASS below.

PASS
